requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
django==5.2.1; python_version >= '3.10'
djangorestframework==3.16.0; python_version >= '3.9'
djangorestframework-simplejwt==5.5.0; python_version >= '3.9'
pyjwt==2.9.0; python_version >= '3.8'
sqlparse==0.5.3; python_version >= '3.8'
